+++ FluidWarfare.Render.Vulkan/Scene3D/Session/VulkanScene3dSession.Frame.cs
@@ -1,4 +1,3 @@
-using System.Runtime.InteropServices;
 using FluidWarfare.Render.Camera;
 using FluidWarfare.Render.Camera.Navigation;
 using FluidWarfare.Render.Vulkan.Camera;
@@ -67,32 +66,5 @@ partial class VulkanScene3dSession
         }
         catch (Exception ex) { System.Diagnostics.Debug.WriteLine($"[Overlay] 异常：{ex.Message}"); }
         return (0, null, null, null);
-    }
-
-    unsafe Result SubmitFrame()
-    {
-        _vk!.GetDeviceQueue(_device, _queueIndex, 0, out _queue);
-        var ws = stackalloc Silk.NET.Vulkan.Semaphore[] { _swapchainRes!.SemAvail };
-        var wst = stackalloc Silk.NET.Vulkan.PipelineStageFlags[] { PipelineStageFlags.ColorAttachmentOutputBit };
-        var cb = stackalloc Silk.NET.Vulkan.CommandBuffer[] { _swapchainRes.CommandBuffer };
-        var ss = stackalloc Silk.NET.Vulkan.Semaphore[] { _swapchainRes.SemFin };
-        var si = new SubmitInfo
-        {
-            SType = StructureType.SubmitInfo, WaitSemaphoreCount = 1, PWaitSemaphores = ws,
-            PWaitDstStageMask = wst, CommandBufferCount = 1, PCommandBuffers = cb,
-            SignalSemaphoreCount = 1, PSignalSemaphores = ss
-        };
-        return _vk.QueueSubmit(_queue, 1, &si, _swapchainRes.Fence);
-    }
-
-    unsafe Result PresentFrame(uint imgIndex)
-    {
-        var fn = Marshal.GetDelegateForFunctionPointer<QueuePresentFn>(_fnQueuePresent);
-        var pws = stackalloc Silk.NET.Vulkan.Semaphore[] { _swapchainRes!.SemFin };
-        var psc = stackalloc SwapchainKHR[] { _swapchainRes.Swapchain };
-        uint idx = imgIndex;
-        var pi = new PresentInfoKHR { SType = StructureType.PresentInfoKhr, WaitSemaphoreCount = 1,
-            PWaitSemaphores = pws, SwapchainCount = 1, PSwapchains = psc, PImageIndices = &idx };
-        return fn(_queue, &pi);
     }
 }